audio-player: Implement dynamic media file selection
This commit implements a dynamic media file selection (playbin uri) based on a playlist file
passed as argument.
This allows the audio-player to play a media file based on the user choice by defining the path
to the media files in a playlist file. The function select_media_file() parses the content of the
playlist and sets the desired media file according to a index. This index will allow to move forward
and backwards between audio media defined in the playlist file.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>