audio-player: Add capability to dinamically update the memory allocation for media file
This commit implements the reallocation of the chunk which stores the path to the media file.
To avoid specifying a maximum media file path length, the selection of the next media file
dinamically resizes the reserved memory when a path is higher that the length defined by default.
Since the selection of the next media file encompasses an iteration through the playlist file, the
memory reserved for the temporary media file is the same throughout executions, if it is expanded once,
the next allocation will employ the new required memory allocation.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>